Workflow Improvements (#109)
This PR includes several changes to improve the automation of the build
and test environment:

### Improvements to GitHub Actions and Workflows:
*
[`.github/actions/prepare_env/action.yaml`](diffhunk://#diff-d2a643aadfc6be30476e47be7a45cfb559795c175427f011ec24e37c91279fb9R1-R50):
Created a new composite action to prepare the environment for the build,
including installing packages, configuring the environment, caching the
toolchain, and building the toolchain if not cached.
*
[`.github/workflows/automated_tests.yaml`](diffhunk://#diff-0aca314923858c03b7dd31233cdc92afe2fa647ff0ab19572de5aeba83b6d1b0L13-R38):
Updated the workflow to use the new `prepare_env` action, build AlkOS,
run tests, and upload artifacts with improved naming and compression.

### Enhancements to Dependency Installation:
*
[`scripts/alkos_cli.bash`](diffhunk://#diff-d6fe9a3e8f641a3f60a80deb12392cfa03d1c044e26fff8f2ad85d8fe5471b73R20-R25):
Refactored the dependency installation process to support multiple
distributions by introducing a dictionary of install scripts and a
function to detect the supported distribution.

### Toolchain Version Management:
*
[`scripts/env/build_cross_compile.bash`](diffhunk://#diff-ffc85218e152db2f3b304662092ffd89c58caf496080efbdff5bec7a0391e7c3R40-R72):
Introduced a function to load toolchain versions from a file, ensuring
that the correct versions are used during the build process.
*
[`scripts/env/toolchain_versions.txt`](diffhunk://#diff-20ff7723555c0c15f487fd6aec7eb7cf21d33b0e1d3d1672badf83949d5e3bb4R1-R3):
Added a new file to specify the versions of binutils, gdb, and gcc to be
used for cross-compilation.

### Cleanup and Removal of Redundant Scripts:
*
[`scripts/cicd/build_and_run_tests.bash`](diffhunk://#diff-3957d4d55ae6838785a506ad14a969ebbb41519142a59e41f17798cdf8914483L1-L21):
Removed the script as its functionality has been integrated into the new
composite GitHub Action and updated workflow.

### Bug Fixes:
* Fixed the progress bar display that was polluting the screen during
toolchain installing.