Create a higher half kernel (#104)
This is a rework wide in scope, that sets to enable a 64-bit higher half
kernel.
In order to do so:
- A loader memory manager is introduced, to allocate physical memory and
map it to the higher half. This is a special type of simplified manager
that can only allocate, and never deallocate, physical memory and manage
virtual memory.
- A split between loader32 and loader64 is made, meaning there are now 2
intermediate stages of loading (loader32->loader64->kernel) this is
required as I can't load the kernel into the higher half of address
space in 32 bit mode with 32 bit addresses. The `loader64` loads the
kernel at the higher half of the virtuall address space and jumps to it.
- The above split neccessitated to create "common-loader-32-64-kernel"
which shares files between all, and "common-loader-64-kernel" which just
shares files between 64 bit parts of x86_64 specific code.
- Further extensions will use the loader memory manager to allocate
enough physical memory for a full stack based physical mem manager in
the kernel and then map it to the an address just above the kernel. This
will be very convenient in the implementation fo a physical memory
manager.
- A set of refactors/cleanups and general improvements to the code were
done
- the kernel and loader64 are now compiled with
`POSITION_INDEPENDENT_CODE ON`, so far this hasn't generated any errors
when applied only to those specific targets, but generates errors when
enabled globally. Further investigation is requiered.
- Appended `CONF_KERNEL_COMMANDS` to the bash configuration file in
`alkos/CMakeLists.txt`. And expanded the make_iso script to append them
in the menuentry. I use this to assosicate each module with a string,
which enables me to differentiate between modules when trying to load
them. (without it loader64 and kernel are indistinguishable from the
perspective of multiboot2 structures)

---------

Co-authored-by: Jakub Lisowski <Jlisowskyy@gmail.com>