Kryczkal/cmake modularization 02 (#159)
# Summary

CMakeLists now includes a target to which to link against. It sets all
the necessary properties.
Idea is -> there is 1 main property interface library for almost
anything.
If arch requires some custom one, it's defined in a .cmake file of the
arch, and the existance of this library is a contract between the arch
CMakeLists.txt and the arch configuration file. I think this is rather
clean.

# Important!
Meddling with this stuff can easily introduce some very subtle bugs if a
property for some reason isn't applied to a target etc. So I hope the
tests will catch if anything like that was introduced. This PR needs to
be approached with care for typos / mistakes etc that could subtly break
this.

# Things done
- Renamed x86_64-flags.cmake to x86_64-conf.cmake as i added the
property libraries here. Maybe split into two files? But I dont feel
there is a need currently.
- Updated all CMakeLists.txt to simply link against those interface
libraries and removed much redundant code.

#158 Link to previous part of the refactor