Kryczkal/decouple libc from kernel (#181)
## Problem
libc is coupled to the kernel which is in turn coupled to loaders. This
makes everything a mess almost impossible to modify without side effects
and random `explosions` (spirals where modification of one thing leads
to cascading chain of errors in almost unrelated stuff).
This is unmaintainable

## Solution
This PR decouples libc from the kernel by introducing abi layer.
- libc is complied as static library with several functions without
provided definitions.
when linking against libc, the target must provide implementations for
the platform specific functions.

## Side Notes
- This will most likely allow a rather painless transition from libk to
libc
- This removes necessity for many akward macro directies like 'ifdef
KERNEL` in libc
- This allows to link libc directly to boot/loaders, without going
through kernel, which led to me further decouple boot (make it a
separate project) in #182