Kryczkal/expected (#184)
## Problem
There are many places in kernel where functions can fail. Exceptions are
disabled. No formal & sane way to handle it.

## Solution

This PR introduces an almost complete (expect a couple stricter
'requirements' checks) implementation of `std::expected`. Every function
that can fail should return `std::expected`. This way it's the callers
responsibility to check if things worked, which provides context for
errors. It's also overall a sane approach.

## Details 
- It follows glibcc very closely. 
- Logic is fundamentally (if you exclude the boilerplate of
'requirements', 'noexcept' and 20 constructors that do more or less the
same) simple.
- It's not tested yet fully. 

##  Side Note
I'll add tests to this shortly when I make the tests work again on this
branch