Kryczkal/kernel pmm (#192)
## Problem
Kernel has no memory management

## Solution
The infra for creatin a proper buddy PMM was created in the loader PR's.
Kernel gets a proper bitmap of all avaliable pages alongside the address
of the pml table. Kernel then has to use it to init a pmm.

### This PR introduced:
- Mixin, that reduces boilerplate for classes that require delayed
initialization.
- basic BitmapPmm,
- skeleton of Vmm (unimplemented) 
- skeleton of a BuddyPmm (unimplemented)
- infrastructure to add them to a MemoryModule, configure them in
PreKernelInit and initialize in KernelInit.
- easy swapping out of Vmm/Pmm implementations with same ABI (ie. Bitmap
for Buddy when it's done)
### In addition to that, in Loader:
- Direct 1:1 Memory Mapping at a specified virtual address (created by
loader).
- PhysicalPtr that abstracts away the logic of accessing physical
addresses via the DirectMemMapping
- Added required variables to be passed from loader to kernel to
actually initialize pmm.
- Separated Loader Vmm definition from Implementation (.hpp and .tpp
split)

## Bugs
Discovered a problem where current HAL folder and abi usage creates a
circular dependency which prevented me from creating multiple PMM's with
same interface. Added a devnote about that.
Also, hal folder has the exact same paths as include folder, so
hal/mem/pmm.hpp and include/mem/pmm.hpp conflict.
I believe the include paths need to be prefixed with hal in the kernel,
and hal/impl/ in the arch (or something like that). There needs to be
easy differenciation between the interface (ABI), the implementation (IN
ARCH), and the file that you use to get the implementation
ie. 
hal/abi/spinlock.hpp - declares interface for spinlock
hal/impl/spinlock.hpp - contains the implementation, is created in arch
hal/spinlock.hpp - imports hal/impl/spinlock.hpp

Three files are needed since the .hpp with abi CANT KNOW ANYTHING ABOUT
IMPLEMENTATION (CYCLIC DEPENDNCY)
but this is out of scope of this PR. To be discussed.