Kryczkal/hal refactor (#198)
## Problem
Architecture api included kernel code, which included architecture code,
which created a cycle, that made implementing kernel abi much more
coupled, and error prone.

## Solution 
Decouple the API (programming interface) from the kernel code that
includes arch specific stuff.

## Other notes
Usage of arch namespace is now responsible for wrapping around
architecture dependent code so that its not accidentally used in kernel.
hal (hardware abstraction layer) namespace on the other hand is the
official interface (bridge) through which to use architecture dependent
code in the kernel.

There were places in kernel where fragile code that was not in the API
was used in core places. This prevents that

---------

Co-authored-by: Jakub Lisowski <Jlisowskyy@gmail.com>