Kryczkal/streams (#223)
### Problem
There is no IO infra. Qemu/Serial/Terminal are not connected in any
unified way, there is no notion of a pipe/stream etc.

### Solution
Implemented a generic interface for Writers / Readers (Consumers and
Producers). Made the notion of a Pipe (Single Producer - Single Consumer
connection via a circular buffer). Created infra for writer that writes
to many sources. Basically set up the infrastructure to allow
redesigning the whole IO in the kernel in future PR's.

### TODO:
Testing - In next PR