Kryczkal/io 01 (#226)
Super small PR for making a generic SerialDriver.
This was created with the intention of integrating the IO Stream API /
MultiplexerWriter into the current Trace framework / print / etc. But I
decided what's already there, can just stay for now. I'm just going to
implement the new Console using this API.
This is the leftover from this idea, before I stopped myself. The lone
generic UART16550 driver (can be used to writing to QEMU)