Kryczkal/console-01 (#228)
## Problem
There is no way to display text on the screen

## Solution
Implement font support in the kernel,
`Font` is a concept of an object that returns `Glyphs`
`Glyphs` are bitmaps of speciifc "characters". These are then handled by
`Painter` to draw them on a `Surfrace`.
- Added support for PSF2 type fonts, so they can be easily added to the
kernel.
 - Added drdos8x8 as a font

 No truetype font support. No fancy stuff. Just plain old bitmasks
 
 
<img width="1384" height="888" alt="image"
src="https://github.com/user-attachments/assets/d2031e9a-ebc4-4f63-9e37-1b782a7cd9a6"
/>
<img width="1290" height="914" alt="image"
src="https://github.com/user-attachments/assets/c3cef99b-65b0-4fef-a449-54671a4f2a57"
/>

## Note
This is not an implementation of a Console yet. A console would be a
`Writer` that you just put text into and don't care about where it's
gonna be on screen, at what position to display it etc.
This is just adding `DrawString` to `Painter` class.

Console is still on the way