Kryczkal/elf loading (#243)
## Problem
Programs need to run in userspace in the future
Programs (elfs) need to be loaded from rootfs
in order to be in rootfs, they need to be compiled, and copied there

<img width="773" height="354" alt="image"
src="https://github.com/user-attachments/assets/ae53dfe3-92fb-431d-b8ae-cb5778438250"
/>


## Solution
This PR introduced:
- the `userspace` folder, that handles creating userspaces programs.
- elf loader that creates proper virtual memory regions (in kernel
address space) and jumps to the loaded program

## Bugfixes
Fixed a bug in VFS Path where non-owning view type (that viewed volatile
mem) objects were copied, without copying the underlying thing they
viewed at, causing errors.

## Note
**TDLR**: Hardcoded for x86_64 and only a single program can be executed
per run of alkos

**Long Version**: This this is severly limited, and more of a Proof of
Concept.
1. There is no cleanup that "unloads" a program (and most likely will
not be needed since each program is a process in its own address space,
so destroying an address space is the cleanup itself). This means a
program can only be run once, then its loaded into memory, and cant be
loaded again (since the vmem areas are already used).
3. Since there is no libc and syscalls yet, this PR added a workaround
where userspace programs are compiled in such a way that address of a
desired print function is passed into them.
This allowed of testing that indeed the elf loading functionality works
out, without waiting for syscalls and libc to be fully implemented. This
needs to be cleaned up later.
This is temporarily hardcoded for `x86_64`